+++ src/app/api/selections/categories/route.ts
@@ -3,6 +3,8 @@ import { getSelectionCategories } from "@/services/selection.services";
 import { NextRequest } from "next/server";
 
 
+export const dynamic = 'force-dynamic';
+
 export const GET = async (request: NextRequest) => {
   try {
     const categories : ISelectionCategory[] = await getSelectionCategories();
