+++ src/app/api/selections/locations/route.ts
@@ -3,6 +3,8 @@ import { getSelectionLocations} from "@/services/selection.services";
 import { NextRequest, NextResponse } from "next/server";
 
 
+export const dynamic = 'force-dynamic';
+
 export const GET = async (request: NextRequest) => {
   try {
     const locations : ISelectionLocation[] = await getSelectionLocations();
