+++ src/app/api/selections/spots/categories/route.ts
@@ -2,6 +2,8 @@ import { getSpotCategories } from '@/services/spot.services';
 import { NextRequest, NextResponse } from 'next/server';
 
 
+export const dynamic = 'force-dynamic';
+
 export const GET = async(request: NextRequest) => {
   const spotCategories = await getSpotCategories();
